Create 03.cpp
Write a program to accept a student’s name and scores in three subject. Display the 1st, 2nd, average and total. Display whether the student has secured 1st , 2nd , pass class or has failed.1st class is for a score of 60 and above, 2nd class is for a score of 50 and above, while pass class is for a score of 35 and above. If the score is less than 35, then the student fails.